pyproject.toml
[project]
name = "fibery-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "click>=8.1.8",
    "httpx>=0.28.1",
    "mcp[cli]>=1.4.1",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-asyncio>=0.25.3",
]
